replace usize with u64
usize doesn't have a stable size, we should use it where it literally doesn't matter what the type is

+++ packages/bits/src/from_bits.rs
@@ -31,7 +31,7 @@ impl FromBits<32> for u32 {
 macro_rules! from_bits_impls {
     ($($n:expr),*) => {
         $(
-            impl FromBits<$n> for usize {
+            impl FromBits<$n> for u64 {
                 fn from_bits(value: impl Bits) -> Self {
                     value.full()
                 }
