+++ packages/gcn/src/instructions/formats/ds.rs
@@ -38,7 +38,7 @@ impl<R: Reader> ParseInstruction<R> for DSInstruction {
     fn parse(token: u32, reader: R) -> Result<Self, anyhow::Error> {
         let token = combine(token, reader)?;
 
-        Ok(DSInstruction::from_bits(token as usize))
+        Ok(DSInstruction::from_bits(token))
     }
 }
